+++ example/lib/main.dart
@@ -18,43 +18,7 @@ class Example extends StatelessWidget {
         title: "SpaceJam example",
         theme: ThemeData(
           primarySwatch: Colors.blue,
-          // TextThemes
-          textTheme: const TextTheme(
-            titleMedium: TextStyle(
-              fontWeight: FontWeight.bold,
-              /* fontSize: (MediaQuery.of(context).size.width +
-                 MediaQuery.of(context).size.height) /
-                 2 * .06,
-              */
-              color: Colors.black,
-            ),
-            titleSmall: TextStyle(
-              /* fontSize: (MediaQuery.of(context).size.width +
-              MediaQuery.of(context).size.height) /
-              2 * .04,
-          */
-              color: Colors.black,
-            ),
-            headlineMedium: TextStyle(
-              //fontSize: MediaQuery.of(context).size.width * .08,
-              fontWeight: FontWeight.bold,
-              color: Colors.black,
-            ),
-            headlineSmall: TextStyle(
-              //fontSize: MediaQuery.of(context).size.width * .05,
-              color: Colors.black,
-              fontWeight: FontWeight.bold,
-            ),
-            bodyMedium: TextStyle(
-              color: Colors.white,
-              fontWeight: FontWeight.bold,
-              //fontSize: MediaQuery.of(context).size.height * .025,
-            ),
-            bodySmall: TextStyle(
-              color: Colors.white70,
-              //fontSize: MediaQuery.of(context).size.height * .02,
-            ),
-          ),
+          textTheme: SpaceJamTheme.textTheme(),
         ),
         home: const HomePage(),
       );
@@ -71,10 +35,126 @@ class HomePage extends StatefulWidget {
 
 /// Stateful part of the Widget.
 class HomePageState extends State<HomePage> {
-  final ScrollController _controller = ScrollController();
-
   @override
-  Widget build(BuildContext context) => Scaffold(
+  Widget build(BuildContext context) => SpaceJamPage(
+        title: "SpaceJam",
+        appBarRightAction: const SpaceJamAppBarAction(Icons.search),
+        children: <Widget>[
+          SpaceJamContainer(
+            title: "SpaceJam",
+            backgroundColor: Colors.blue,
+            children: <Widget>[
+              Padding(
+                padding: EdgeInsets.only(
+                  left: (MediaQuery.of(context).size.width +
+                          MediaQuery.of(context).size.height) /
+                      2 *
+                      .02,
+                  right: (MediaQuery.of(context).size.width +
+                          MediaQuery.of(context).size.height) /
+                      2 *
+                      .02,
+                ),
+                child: Column(
+                  crossAxisAlignment: CrossAxisAlignment.start,
+                  children: <Widget>[
+                    Text(
+                      "Title",
+                      style: SpaceJamTextStyles.title(context),
+                    ),
+                    Text(
+                      "Subtitle",
+                      style: SpaceJamTextStyles.titleSmall(context),
+                    ),
+                    Text(
+                      "Headline",
+                      style: SpaceJamTextStyles.headline(context),
+                    ),
+                    Text(
+                      "Subheading",
+                      style: SpaceJamTextStyles.headlineSmall(context),
+                    ),
+                    Text(
+                      "Body",
+                      style: SpaceJamTextStyles.bodyMedium(
+                        context,
+                        color: Colors.black,
+                      ),
+                    ),
+                    Text(
+                      "Caption",
+                      style: SpaceJamTextStyles.bodySmall(
+                        context,
+                        color: Colors.black,
+                      ),
+                    ),
+                  ],
+                ),
+              ),
+              const SizedBox(
+                height: 10,
+              ),
+              SpaceJamButton(
+                valueFontSize: 32,
+                titleFontSize: 24,
+                title: "Test",
+                value: "ImageViewer",
+                background: Colors.white24,
+                action: () {
+                  Navigator.push(
+                    context,
+                    MaterialPageRoute<Widget>(
+                      builder: (BuildContext context) => SpaceJamImagePage(
+                        locale: const Locale("en"),
+                        image: Image.network(
+                          "https://assets.4cdn.hu/kraken/6pI26jBO69hwv9c6s.jpeg",
+                        ),
+                        imageURL:
+                            "https://assets.4cdn.hu/kraken/6pI26jBO69hwv9c6s.jpeg",
+                      ),
+                    ),
+                  );
+                },
+              ),
+            ],
+          ),
+          SpaceJamCollection(
+            title: "Collection",
+            items: <SpaceJamCollectionItem>[
+              const SpaceJamCollectionItem(
+                backgroundColor: Colors.blue,
+                upperCaption: "caption",
+                upperValue: "value",
+                lowerCaption: "caption",
+                lowerValue: "value",
+              ),
+              const SpaceJamCollectionItem(
+                backgroundColor: Colors.blue,
+                upperCaption: "caption",
+                upperValue: "value",
+                lowerCaption: "caption",
+                lowerValue: "value",
+              ),
+              const SpaceJamCollectionItem(
+                backgroundColor: Colors.blue,
+                upperCaption: "caption",
+                upperValue: "value",
+                lowerCaption: "caption",
+                lowerValue: "value",
+              ),
+              SpaceJamCollectionItem(
+                backgroundColor: Colors.blue,
+                onTap: () {},
+                upperCaption: "caption",
+                upperValue: "value",
+                lowerCaption: "caption",
+                lowerValue: "value",
+              ),
+            ],
+          ),
+        ],
+      );
+  /*Scaffold(
         backgroundColor: Colors.white,
         body: Stack(
           children: <Widget>[
@@ -175,9 +255,6 @@ class HomePageState extends State<HomePage> {
                         ),
                       ],
                     ),
-                    const SizedBox(
-                      height: 0,
-                    ),
                     SpaceJamCollection(
                       title: "Collection",
                       items: <SpaceJamCollectionItem>[
@@ -225,5 +302,5 @@ class HomePageState extends State<HomePage> {
             ),
           ],
         ),
-      );
+      );*/
 }