+++ lib/spacejam.dart
@@ -1,6 +1,8 @@
 export "src/theme/text_styles.dart";
+export "src/theme/theme.dart";
 
 // Widgets
+export "src/page/spacejam_page.dart";
 export "src/collection/collection_item.dart";
 export "src/container/container.dart";
 export "src/collection/collection.dart";