+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 0.3.0
 * Added Content
 * Added ContentItem
+* Added SpaceJamPage
 * Added `duration` property to Appbar
 
 #### Breaking changes:
@@ -8,10 +9,13 @@ Appbar:
 - Renamed `Appbar` to `SpaceJamAppBar`. 
 AppBarAction:
 - Renamed `AppBarAction` to `SpaceJamAppBarAction`.
+- `icon` is now an unnamed parameter.
 Button:
 - Renamed `Button` to `SpaceJamButton`.
 ContentBox:
 - Renamed `ContentBox` to `SpaceJamContainer`.
+ImagePage:
+- Renamed `ImagePage` to `SpaceJamImagePage`.
 # 0.2.2
 * Bugfixes
 # 0.2.1